requirements.txt
gitdb==4.0.12
GitPython==3.1.44
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
shiboken6==6.9.1
smmap==5.0.2
